requirements.txt
# #############################################################################
# requirements.txt
# ================
# Author : Sepand KASHANI [kashani.sepand@gmail.com]
# #############################################################################

# This file contains the complete list of Pypeline's Python dependencies.  It is preferable to
# install packages having non-Python dependencies using the `conda` environment manager if available.
#
# When modifying dependencies, don't forget to update both `requirements.txt` and
# `conda_requirements.txt`.

astropy == 4.0.*
black == 19.3b0
imot_tools == 1.0
ipython == 7.5.*
matplotlib == 3.1.*
numexpr == 2.6.*
numpy == 1.16.*
pandas == 0.24.*
plotly == 3.5.*
pyffs == 1.0
pyproj == 1.9.*
pytest == 4.5.*
python-casacore == 2.2.*
scikit-learn == 0.21.*
scipy == 1.2.*
sphinx == 1.8.*
sphinx_rtd_theme == 0.4.*
tqdm == 4.29.*
